requirements.txt
absl-py==1.0.0
alabaster==0.7.12
alembic==1.7.5
appdirs==1.4.4
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asn1crypto==1.4.0
astor==0.8.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
autopage==0.4.0
babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.0
bitstring==3.1.9
blist==1.3.6
# box2d==2.3.2
cachecontrol==0.12.6
cached-property==1.5.2
cachetools==5.0.0
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.7
cliff==3.10.0
clikit==0.6.2
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.4
colorlog==6.6.0
crashtest==0.3.1
cryptography==3.4.7
cycler==0.10.0
cython==0.29.24
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
ecdsa==0.17.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==2.0
flit-core==3.3.0
flit==3.3.0
fsspec==2021.7.0
future==0.18.3
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.11
gitpython==3.1.26
glfw==2.6.4
glob2==0.7
google-auth-oauthlib==0.4.6
google-auth==2.6.0
google-pasta==0.2.0
graphviz==0.14
greenlet==1.1.2
grpcio==1.44.0
#gym-notices==0.0.8
gym==0.22.0
h5py==3.6.0
html5lib==1.1
idna==3.3
imageio-ffmpeg==0.4.7
imageio==2.13.5
imagesize==1.2.0
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
intervaltree==3.1.0
intreehooks==1.0
ipaddress==1.0.23
ipykernel==6.7.0
ipython-genutils==0.2.0
ipython==7.31.1
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.0.3
joblib==1.0.1
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras-preprocessing==1.1.2
keras==2.7.0
keyring==21.2.0
keyrings.alt==4.1.0
kiwisolver==1.3.1
labmaze==1.0.5
liac-arff==2.5.0
libclang==13.0.0
lockfile==0.12.2
lxml==4.8.0
mako==1.1.6
markdown==3.3.6
markupsafe==2.1.3
matplotlib-inline==0.1.3
matplotlib==3.4.0
mazelib==0.9.14
mistune==0.8.4
mock==4.0.3
more-itertools==8.8.0
moviepy==1.0.3
msgpack==1.0.2
mujoco==2.1.5
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
nose==1.3.7
notebook==6.4.7
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
optuna==2.10.0
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.7.2
parso==0.8.3
pastel==0.2.1
pathlib2==2.3.6
pathtools==0.1.2
pbr==5.8.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.0.0
pip==21.2.2
pkginfo==1.7.1
platformdirs==2.2.0
pluggy==0.13.1
poetry-core==1.0.3
poetry==1.1.7
portpicker==1.5.0
prettytable==3.0.0
proglog==0.1.10
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
py-expression-eval==0.3.13
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pydot==1.4.1
pyglet==1.5.21
pygments==2.11.2
pylev==1.4.0
pynacl==1.4.0
pyopengl==3.1.6
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.1
pytest==6.2.4
python-dateutil==2.8.2
python-editor==1.0.4
pytoml==0.1.21
pytz==2023.3.post1
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.0
regex==2021.8.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests==2.27.1
rsa==4.8
scandir==1.10.0
scipy==1.7.3
seaborn==0.12.2
secretstorage==3.3.1
semantic-version==2.8.5
send2trash==1.8.0
sentry-sdk==1.10.1
setproctitle==1.3.2
setuptools-rust==0.12.1
setuptools-scm==6.0.1
setuptools==69.0.3
shellingham==1.4.0
shortuuid==1.0.9
simplegeneric==0.8.1
simplejson==3.17.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
sqlalchemy==1.4.30
stevedore==3.5.0
tabulate==0.8.9
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.15.0
tensorflow==2.7.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-agents==0.11.0
threadpoolctl==2.2.0
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.1.1
tzdata==2023.4
ujson==4.0.2
urllib3==1.26.18
virtualenv==20.7.0
wandb==0.16.2
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==2.0.3
wheel==0.42.0
widgetsnbextension==3.5.2
wrapt==1.13.3
xlrd==2.0.1
zipfile36==0.1.3
zipp==3.7.0
